Using double in Rectangle

+++ src/com/tw/step8/assignment1/Rectangle.java
@@ -1,18 +1,18 @@
 package com.tw.step8.assignment1;
 
 public class Rectangle {
-  private final int length;
-  private final int breadth;
+  private final double length;
+  private final double breadth;
 
-  public Rectangle(int length, int breadth) {
+  public Rectangle(double length, double breadth) {
     this.length = length;
     this.breadth = breadth;
   }
-  public int area() {
+  public double area() {
      return this.length * this.breadth;
   }
 
-  public int perimeter() {
+  public double perimeter() {
     return 2 * (this.length + this.breadth);
   }
 }
